fix integration test regression
When using a configuration class derived from XtextConfiguration, as opposed to using that class directly, Guice injection failed, because the GuiceBundle adds a binding to the exact type of the configuration object, only. This means that was no explicit binding for the base class XtextConfiguration in that situation, causing Guice to default to creating an instance using the default constructor. The problem is circumvented by adding an explicit binding for the base type, unless the configuration class is not a subtype, in which case the default behavior is sufficient (and would actually clash with adding another explicit binding for the same type).