fix issue finding ProjectDescription
When resources are loaded during the initial build of the index,
resource(set)s do not use CustomWebResourceSetProvider, thereby
skipping the project description's registration with the resource
set. Therefore, the ChunkedResourceDescriptionsProvider does not
assume that resource sets being passed in are already registered
with its project description, but rather checks this itself and
performs the registration itself, if necessary.