refactor in response to pull request #30 remarks
Simplified ChangeDetector API by removing the
chain of responsibility pattern and replacing it
with a simple class that depends directly on the
required detectors. While both less generic and
less configurable, it considerably reduces API
complexity, while still retaining a level of
flexibility that should be sufficient for current
requirements.