add test to reproduce bug, fix by binding different IContainer.Manager.
XtextIndexModule now binds SimpleResourceDescriptionsBasedContainerManager,
instead of ProjectDescriptionBasedContainerManager. The former delegates to
the latter automatically, if appropriate, i.e. if the IResourceDescriptions
instance being passed to its method(s) is a ChunkedResourceDescriptions
instance. Using ProjectDescriptionBasedContainerManager directly had caused
an exception when an instance of ResourceDescriptionsData was passed in,
instead. IncrementalBuilder / Indexer use instances of that type by way of
IndexState.